+++ src/builder.lisp
@@ -1,8 +1,10 @@
 (defpackage #:ultralisp/builder
   (:use #:cl)
+  (:import-from #:ultralisp/clpi)
   (:import-from #:local-time
                 #:now
                 #:format-timestring)
+  (:import-from #:ultralisp/clpi)
   (:import-from #:quickdist
                 #:render-template
                 #:quickdist)
@@ -69,12 +71,8 @@
                 #:source-release-info)
   (:import-from #:ultralisp/models/dist-source
                 #:dist->sources)
-  (:export
-   #:build
-   #:build-version
-   #:build-prepared-versions
-   #:prepare-pending-version
-   #:build-pending-dists))
+  (:export #:build
+           #:build-pending-dists))
 (in-package ultralisp/builder)
 
 
@@ -108,173 +106,6 @@
              :projects-dir projects-dir
              :dists-dir dist-dir
              :version (get-new-version-number)))
-
-
-(defclass save-version-command ()
-  ((version :initarg :version
-            :type version
-            :reader get-version)))
-(defun make-save-version-command (version)
-  (check-type version version)
-  (make-instance 'save-version-command :version version))
-
-
-(defclass upload-command ()
-  ((dir :initarg :dir
-        :reader get-dir)))
-
-
-(defun make-upload-command (dir)
-  (make-instance 'upload-command :dir dir))
-
-
-(defclass disable-project-command ()
-  ((project :initarg :project
-            :type project-version
-            :reader get-project)
-   (reason :initarg :reason
-           :type keyword
-           :reader get-reason)
-   (description :initarg :description
-                :type string
-                :reader get-description)))
-
-
-(defun make-disable-project-command (project reason description)
-  (check-type project project-version)
-  (check-type reason keyword)
-  (check-type description string)
-  (make-instance 'disable-project-command
-                 :project project
-                 :reason reason
-                 :description description))
-
-
-(defgeneric perform (command)
-  (:documentation "This method is called by the main process
-                   to change state of the database. Because worker
-                   is unable to do it itself."))
-
-
-(defmethod perform ((command save-version-command))
-  (save-dao (get-version command)))
-
-
-(defmethod perform ((command disable-project-command))
-  (let* ((project-version (get-project command))
-         (reason (get-reason command))
-         (description (get-description command))
-         (project (ultralisp/models/project:get-project project-version)))
-    (log:error "Disabling project" project reason description)
-    (disable-project project
-                     :reason reason
-                     :traceback description)))
-
-
-(defmethod perform ((command upload-command))
-  (let* ((dir (get-dir command)))
-    (log:info "Uploading" dir)
-    (error "This code was broken at some moment, upload requires two parameters. But everything seems work. Probably it is not executed.")
-    ;; (upload dir)
-    ))
-
-
-;; TODO: remove this function and probably all commands above
-(defun build-version-remotely (version
-                               &key
-                                 (projects-dir (get-projects-dir))
-                                 (name (get-dist-name))
-                                 (base-url (get-base-url))
-                                 (dist-dir (get-dist-dir))
-                                 db-user
-                                 db-pass
-                                 db-host
-                                 (download-p t))
-  "This function will be performed on a worker with
-   read-only access to the database.
-
-   It should return a list of commands to the main process to
-   modify the database."
-  (with-fields (:request-id (make-request-id))
-    (check-type version version)
-    
-    (handler-bind ((error (lambda (condition)
-                            ;; We want debugger to popup if we've connected to
-                            ;; the process from SLY
-                            ;; (invoke-debugger condition)
-                            (when ultralisp/slynk:*connections*
-                              (invoke-debugger condition)))))
-      (with-log-unhandled ()
-        (log:info "Building a new version" version)
-    
-        (with-connection (:username db-user
-                          :password db-pass
-                          :host db-host)
-          (uiop:ensure-all-directories-exist (list projects-dir))
-          
-          (let* ((projects-dir (truename* projects-dir))
-                 (_ (when download-p
-                      (log:info "Downloading projects")))
-                 (downloaded-projects (when download-p
-                                        (download version projects-dir)))
-                 (num-downloaded-projects (length downloaded-projects))
-                 commands)
-            (declare (ignorable _))
-            
-            (when download-p
-              (log:info "Projects were downloaded" num-downloaded-projects)
-
-              (log:info "Removing disabled projects from disk")
-              (remove-disabled-projects projects-dir
-                                        downloaded-projects))
-
-            (handler-bind ((error (lambda (condition)
-                                    (let ((restart (find-restart 'quickdist:skip-project)))
-                                      (cond
-                                        (restart
-                                         (log:error "Error catched during processing" quickdist:*project-path* condition)
-                                         (let ((project (find-project-by-path downloaded-projects
-                                                                              quickdist:*project-path*)))
-                                           (log:info "Sending back command which will disable project" project)
-                                           (push (make-disable-project-command project
-                                                                               :build-error
-                                                                               (print-backtrace :condition condition
-                                                                                                :stream nil))
-                                                 commands))
-                                         (invoke-restart restart))
-                                        (t (error "No skip-project restart found!")))))))
-
-              (let ((version-number (make-version-number)))
-                (setf (get-number version)
-                      version-number)
-                
-                (log:info "Starting quickdist build" version-number)
-                (quickdist :name name
-                           :base-url base-url
-                           :projects-dir projects-dir
-                           :dists-dir dist-dir
-                           :version version-number)))
-            
-            (setf (get-built-at version)
-                  (local-time:now)
-                  (get-type version)
-                  :ready)
-
-            ;; TODO: probably it is not the best idea to upload dist-dir
-            ;;       every time, because there can be previously built distributions
-            ;;       May be we need to minimize network traffic here and upload
-            ;;       only a part of it or make a selective upload which will not
-            ;;       transfer files which already on the S3.
-            (log:info "Sending back command which will upload distribution")
-            (push (make-upload-command dist-dir)
-                  commands)
-            ;; Here we don't save version object because
-            ;; this function will be called on a remote worker
-            ;; without "write" access to the database.
-            (log:info "Sending back command which will save version to the database")
-            (push (make-save-version-command version)
-                  commands)
-            commands))))))
 
 
 (defun create-metadata-for (dist path &key (version-number (dist-quicklisp-version dist)))
@@ -339,49 +170,6 @@
                              :if-exists :append))))))
 
 
-;; TODO: remove after migraion
-(defun build-version (version)
-  "This function generates all necessary metadata for the version and uploads them to the server."
-  (check-type version version)
-  (ultralisp/utils:with-tmp-directory (path)
-    (log:info "Building a new" version "in the" path)
-  
-    (let ((version-number (make-version-number)))
-      (setf (get-built-at version) (local-time:now)
-            (get-type version) :ready
-            (get-number version) version-number)
-    
-      (create-metadata-for version path)
-      (upload path "/")
-      (save-dao version))))
-
-
-;; TODO: remove after migration 
-(defun prepare-pending-version ()
-  (with-transaction
-    (log:info "Trying to acquire a lock performing-pending-checks-or-version-build from prepare pending version to prepare pending version 1")
-    (with-lock ("performing-pending-checks-or-version-build")
-      (let ((version (get-pending-version)))
-        (when version
-          (log:info "Preparing version for build" version)
-          (create-projects-snapshots-for version)
-          (setf (get-type version)
-                :prepared)
-          (save-dao version))))))
-
-;; TODO: remove after migration 
-(defun build-prepared-versions ()
-  "Searches and builds a pending version if any."
-  (with-transaction
-    (log:info "Trying to acquire a lock performing-pending-checks-or-version-build to build prepare version 1")
-    (with-lock ("performing-pending-checks-or-version-build")
-      (log:info "Checking if there is a version to build")
-      
-      (loop for version in (get-prepared-versions)
-            do (build-version version)))))
-
-
-
 (defun prepare-dist (dist)
   (check-type dist ultralisp/models/dist:dist)
   
@@ -400,7 +188,7 @@
                           ;; This added source have to be checked first.
                           unless (eql disable-reason-type
                                       :just-added)
-                          collect source)))
+                            collect source)))
     
       (when changes
         (let ((version-number (make-version-number)))
@@ -419,7 +207,12 @@
                             ;; but sometimes can be useful for debugging:
                             :ready))
       (error "Unable to build dist with state ~S" state)))
-  
+
+  ;; First, we'll update CLPI metadata
+  (log:info "Updating CLPI index for" dist)
+  (ultralisp/clpi::write-index-for-dist dist)
+
+  ;; And then generate metadata in the Quicklisp format
   (ultralisp/utils:with-tmp-directory (path)
     (log:info "Building the" dist "in the" path)
      
@@ -427,7 +220,7 @@
           (dist-state dist) :ready)
 
     (create-metadata-for dist path)
-    (upload path "/")
+    (upload path :quicklisp "/")
     ;; NOTE: in case of error in the database
     ;; we might end with situation when the release
     ;; was uploaded but database not updated.