+++ src/uploader/fake.lisp
@@ -2,20 +2,27 @@
   (:use #:cl)
   (:import-from #:log4cl)
   (:import-from #:ultralisp/uploader/base
-                #:make-uploader)
+                #:make-uploader
+                #:make-files-inclusion-checker)
   (:import-from #:ultralisp/variables
-                #:get-dist-dir)
+                #:get-dist-dir
+                #:get-clpi-dist-dir)
   (:import-from #:metatilities
                 #:relative-pathname))
 (in-package ultralisp/uploader/fake)
 
 
-(defmethod make-uploader ((type (eql :fake)))
-  (lambda (dir-or-file destination-path)
-    (let ((destination-path (relative-pathname (get-dist-dir)
-                                               destination-path)))
+(defmethod make-uploader ((type (eql :fake)) repo-type)
+  (lambda (dir-or-file destination-path &key only-files)
+    (let* ((destination-path (relative-pathname (ecase repo-type
+                                                  (:quicklisp (get-dist-dir))
+                                                  (:clpi (get-clpi-dist-dir)))
+                                                destination-path))
+           (need-to-upload-p
+             (make-files-inclusion-checker only-files)))
       (ultralisp/utils:walk-dir (dir-or-file absolute relative)
-        (let ((destination (merge-pathnames relative destination-path)))
-          (log:info "Copying" absolute "to" destination)
-          (ensure-directories-exist destination)
-          (uiop:copy-file absolute destination))))))
+        (when (funcall need-to-upload-p relative)
+          (let ((destination (merge-pathnames relative destination-path)))
+            (log:info "Copying" absolute "to" destination)
+            (ensure-directories-exist destination)
+            (uiop:copy-file absolute destination)))))))